pyproject.toml
[project]
name = "yolo-training"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "siddhi47", email = "siddhi.47.skb@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "clearml>=2.0.2",
    "hydra-core>=1.3.2",
    "ultralytics>=8.3.175",
    "torch==2.8.0+cu126",
    "torchvision==0.23.0+cu126",
    "boto3>=1.35.0",
    "pyyaml>=6.0",
    "opencv-python>=4.10.0",
    "tqdm>=4.66.0",
]

[project.scripts]
yolo-training = "yolo_training:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
